feat: Results 예외 처리
